+++ CHANGES.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- Upgrade Lindera 0.8.0 #17 @johtani
+
 ## 0.7.1 (2020-10-15)
 - Bump up version to 0.7.1 #15 @mosuka 
 - Automate release tasks #14 @mosuka